okay okay, little late to change the samplet... now, I can't afford new change because the time!
basically the idea of use Lost Universe like a cenario died. But, I'm will keeping using the names of actors (or characters). Basically, it's a normal house with normal people doing what do normally.

AnnotatedEnvironment.java: This class has removing because I don't like it. But, I keep the name of it to use on the new class.